% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_ze_sat.R
\name{compute_Ze}
\alias{compute_Ze}
\title{Compute the euphotic depth from surface satellite Chlorophyll-a concentration}
\usage{
compute_Ze(CHLsat)
}
\arguments{
\item{CHLsat}{the value of satellite Chl (mg.m-3)}
}
\value{
The function returns the value of the euphotic depth (m)
}
\description{
Compute the euphotic depth from surface satellite Chlorophyll-a concentration
}
\examples{
compute_Ze(0.1)

}
\author{
Raphaëlle Sauzède \email{raphaelle.sauzede@imev-mer.fr}
}
